split up generic segmented telescope classes and LUVOIR classes
